Print number

+++ 0x04-more_functions_nested_loops/101-print_number.c
@@ -20,5 +20,5 @@ void print_number(int n)
 
 	if (num != 0)
 		print_number(num);
-	_putchar('0' + num % 10);
+	_putchar('0' + (unsigned int) num % 10);
 }
